+++ lib/spheric/game/behaviors/mixer.ex
@@ -12,6 +12,7 @@ defmodule Spheric.Game.Behaviors.Mixer do
     recipes: [
       %{inputs: [crude_oil: 1, titanium_ingot: 1], output: {:catalysed_fuel, 1}},
       %{inputs: [catalysed_fuel: 8, refined_fuel: 8], output: {:unstable_fuel, 16}},
-      %{inputs: [unstable_fuel: 5, sulfur_compound: 1], output: {:stable_fuel, 2}}
+      %{inputs: [unstable_fuel: 5, sulfur_compound: 1], output: {:stable_fuel, 2}},
+      %{inputs: [water: 10, titanium_dust: 2], output: {:thermal_slurry, 10}}
     ]
 end