+++ lib/spheric/game/behaviors/smelter.ex
@@ -12,6 +12,7 @@ defmodule Spheric.Game.Behaviors.Smelter do
       %{inputs: [iron_ore: 1], output: {:iron_ingot, 1}},
       %{inputs: [copper_ore: 1], output: {:copper_ingot, 1}},
       %{inputs: [titanium_ore: 1], output: {:titanium_ingot, 1}},
-      %{inputs: [raw_quartz: 2], output: {:quartz_crystal, 1}}
+      %{inputs: [raw_quartz: 2], output: {:quartz_crystal, 1}},
+      %{inputs: [ice: 1], output: {:water, 1}}
     ]
 end